Added ABM analysis for gender differences in STEM
This analysis is the one, that I am planing to present on Python Meetup in Ljubljana in december 2018 as a lightning talk. It is longer than what the lightning talk requires.

The analysis shows, how the gender differences depend also on gender differences in intelligence/skill, interestes and habits. It uses agent based modeling as a method of analysis.

I have created this analysis to show, that culturally or even worse oppresive cultures are not the only reason, why they are not more women in STEM, or asking more specifically, why they are not more women programmers.